reordered projects so that it is using columns instead of flex
